Refactor cmd_search.c and add unit tc
